+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ScenarioRunner logs Excel input file name
 - utils.df_itertuples_with_index_names()
 - Core02ScenarioGenerator and Core02ScenarioConfig: configure LexOptiLevels and LexOptiGoals
+- Core01DataManager.remove_zero_quantity_output to filter rows from a df where a solution column is less than the threshold
 ### Fixed
 - Small error in ScenarioRunner.insert_outputs_in_db